refactor(vcvalidator): convert middleware to pipeline Step, consolidate sources
Rejections now flow through the handler's standard signed-NACK path
instead of a hand-rolled unsigned NACK envelope: the plugin implements
definition.StepProvider, runs in the module's steps pipeline, and
returns model.BadReqErr (malformed credential) or model.SignValidationErr
(proof/issuer/expiry/revocation/resolution failure). The machine-readable
failure class (e.g. CREDENTIAL_REVOKED) is preserved at the start of the
NACK error message. The middleware wiring, nackResponse struct and
per-class HTTP status mapping are gone.

Sources consolidated to two files matching sibling plugin layout:
vcvalidator.go (Step, Config, credential extraction) and verify.go
(verification engine: proof/JWT checks, DID resolution, revocation,
base58). Logging switched from fmt.Printf to pkg/log.

Step-level tests cover pass-through (disabled, non-gated action, no
credentials, valid credential) and the rejection-to-model-error mapping.
